refactor: split invoice into connector and app invoice-type #389
